Store edge width as i32

+++ src/popups.rs
@@ -1,7 +1,7 @@
 pub struct Popups {
     // Const Width Constraint popup fields
     const_width_constraint_popup_id: egui::Id,
-    const_width_constraint_user_input: f32,
+    const_width_constraint_user_input: i32,
     const_width_constraint_submitted: bool,
 }
 
@@ -9,7 +9,7 @@ impl Popups {
     pub fn open_const_width_constraint_popup_below_widget(
         &mut self,
         ui: &mut egui::Ui,
-        intial_width: f32,
+        intial_width: i32,
     ) {
         ui.memory_mut(|mem| mem.toggle_popup(self.const_width_constraint_popup_id));
         self.const_width_constraint_user_input = intial_width;
@@ -44,7 +44,7 @@ impl Popups {
         self.const_width_constraint_submitted
     }
 
-    pub fn const_width_constraint_user_input(&self) -> f32 {
+    pub fn const_width_constraint_user_input(&self) -> i32 {
         self.const_width_constraint_user_input
     }
 
@@ -57,7 +57,7 @@ impl Default for Popups {
     fn default() -> Self {
         Self {
             const_width_constraint_popup_id: "const_width_constraint_popup_id".into(),
-            const_width_constraint_user_input: 0.0,
+            const_width_constraint_user_input: 0,
             const_width_constraint_submitted: false,
         }
     }
